requirements.txt
numpy>=1.24
Pillow>=10.0
scipy>=1.10
scikit-image>=0.21
